requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
langchain
langchain-community
langchain-groq==0.3.4
sentence-transformers
faiss-cpu
